requirements_win.txt
absl-py==0.2.2
astor==0.6.2
astroid==1.6.5
bleach==1.5.0
certifi==2018.4.16
cffi==1.11.5
colorama==0.3.9
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
isort==4.3.4
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
numpy==1.14.5
Pillow==5.1.0
protobuf==3.6.0
pycparser==2.18
pylint==1.9.2
PyYAML==3.12
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.10.11
